requirements.txt
django==5.1.8
pymysql>=1.1
Pillow>=10.0
django-cors-headers>=4.0
djangorestframework>=3.15
python-dotenv>=1.0
openai>=1.0
gTTS>=2.5
requests>=2.31
beautifulsoup4>=4.12
# For Gemini:
google-generativeai>=0.5
django-rosetta>=0.10

# ── PDF Processing ────────────────────────────────────────────────────────────
pymupdf>=1.23            # PyMuPDF (fitz) — fast PDF text extraction
pdfplumber>=0.10         # pdfplumber — structured text & table extraction

# ── Word Processing ───────────────────────────────────────────────────────────
python-docx>=1.1         # Word → plain text
mammoth>=1.6             # Word → clean HTML / raw text

# ── OCR (Scanned PDFs) ────────────────────────────────────────────────────────
# Uncomment ONE of the following depending on your environment:
# python-doctr[torch]>=0.8       # docTR with PyTorch backend (CPU/GPU)
# python-doctr[tensorflow]>=0.8  # docTR with TensorFlow backend

# ── Unstructured Content ──────────────────────────────────────────────────────
# Uncomment to enable mixed-content extraction:
# unstructured[pdf]>=0.12